Fix an issue where a parameter without a set value was being reported as not existing.
A parameter can exist in CloudTruth without an explicitly configured value. I.e., there is a difference between a parameter without a value set and a parameter with an empty value. The CLI client had been incorrectly treating the former case as a missing parameter. Now the states are properly accounted for.